migrate unit test for grpc_api
