+++ test/unit/test_otaclient/test_grpc/test_api_v2/conftest.py
@@ -1,0 +1,45 @@
+# Copyright 2022 TIER IV, INC. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import pytest
+
+from otaclient.configs._ecu_info import ECUInfo, parse_ecu_info
+
+ECU_INFO_YAML = """\
+format_version: 1
+ecu_id: "autoware"
+ip_addr: "10.0.0.1"
+bootloader: "grub"
+secondaries:
+    - ecu_id: "p1"
+      ip_addr: "10.0.0.11"
+    - ecu_id: "p2"
+      ip_addr: "10.0.0.12"
+available_ecu_ids:
+    - "autoware"
+    - "p1"
+    - "p2"
+"""
+
+
+@pytest.fixture
+def ecu_info_fixture(tmp_path: Path) -> ECUInfo:
+    _yaml_f = tmp_path / "ecu_info.yaml"
+    _yaml_f.write_text(ECU_INFO_YAML)
+    _, res = parse_ecu_info(_yaml_f)
+    return res
